% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{extract_date}
\alias{extract_date}
\title{Extracts date from a string}
\usage{
extract_date(char, format = c("\%d", "\%m", "\%y"), sep = "-")
}
\arguments{
\item{char}{a character vector with dates to be extracted}

\item{format}{= c("\%d","\%m","\%y")}

\item{sep}{="-" separator}
}
\value{
a character vector
}
\description{
Extracts date from a string
}
