% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{fix_colnames}
\alias{fix_colnames}
\title{Title fix colnames of data frame}
\usage{
fix_colnames(df, pattern)
}
\arguments{
\item{df}{data frame}

\item{pattern}{a pattern to be removed from column names}
}
\value{
data.frame with corrected column names
}
\description{
Title fix colnames of data frame
}
\examples{
# df \%>\% fix_colnames("prefix")
}
