% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{generate_multiple_vars}
\alias{generate_multiple_vars}
\title{generate_multiple_vars}
\usage{
generate_multiple_vars(col, .fun, vec, name, sep = "_", ...)
}
\arguments{
\item{col}{\itemize{
\item vectoor to transform
}}

\item{.fun}{\itemize{
\item R function which takes as argument a col and one of values provided in vec
}}

\item{vec}{\itemize{
\item vector with arguments to use in
}}

\item{name}{\itemize{
\item new column column name
}}

\item{sep}{\itemize{
\item separator to be useed in new column names
}}

\item{...}{\itemize{
\item additional arguments to be passed to .fun
}}
}
\value{
data.frame
}
\description{
generate_multiple_vars
}
\examples{
#generate_new_vars(c(1,10),lag,c(2,3,4),"lagged_columns","_")
}
