% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{load_packages}
\alias{load_packages}
\title{loads packages and attempt to install in not installed}
\usage{
load_packages(packages, verbose = T)
}
\arguments{
\item{packages}{a character vector with packkages names to load}
}
\value{
data.frame with corrected column names
}
\description{
loads packages and attempt to install in not installed
}
